feat(db): SQLite feature Phase B.3 - the compose (SQLite-less base + archive runs)
The payoff of Phase B: a SQLite-less DB-core base + libhull_feature-sqlite.a,
joined by `hull build --with=sqlite`, runs a real db.query through the composed
backend. The base carries ZERO sqlite; the produced app gets the engine +
backend from the archive. No change to the default build.

- Split agent/helpers.c: hl_agent_open_app_db (raw sqlite3 opener) moves to the
  new agent/db_open.c so it travels into the archive; helpers.c keeps the
  backend-agnostic hl_agent_write_error / hl_agent_detect_entry.
- Makefile: FEATURE_SQLITE_OBJS gains cap_db.o (the SQLite stmt-cache / query
  engine that defines hl_cap_db_query / hl_stmt_cache_* -- the missing link
  symbols) and agent_db_open.o. Filter the SQLite-only agent TUs (agent/db.c,
  sql.c, schema_diff.c, db_open.c) out of the base when HL_ENABLE_SQLITE!=1
  (they ship in the archive; the base keeps agent/db_stub.c's A.2 weak stubs).
- build.lua: a `sqlite` FEATURE_SPECS entry (backend hl_db_backend_sqlite, hook
  hl_db_feature_backends, base_group = true) so the compose generates the strong
  hl_db_feature_backends override and wraps platform lib + archive in a
  --start-group for the base<->archive cross-refs. Slots in beside
  postgres/mysql/duckdb (all merge into one override).

Validated end to end by tests/e2e_feature_sqlite.sh (+ `make e2e-feature-sqlite`):
builds the archive (SQLite=1) + a SQLite-less base (HL_SQLITE_FEATURE=1: nm shows
0 sqlite3_open, DB core intact), composes `hull build --with=sqlite`, and runs a
db.query returning 42 through the composed backend (the base had no backend). The
app binary carries sqlite from the archive. Default `make` + `make test` 60/60
unaffected; the default base still has SQLite; agent db schema still works.

Remaining: B.4 -- the toolchain force-load so a SQLite-less `hull` itself runs
`hull test` (:memory:) + the agent DB commands, and Phase C/D (needs_sqlite
auto-compose + embed/publish).
